+++ data/remote_data_source/vod/src/main/java/com/sanaa/vod/interceptor/LanguageInterceptor.kt
@@ -1,0 +1,29 @@
+package com.sanaa.vod.interceptor
+
+import com.sanaa.preferences.service.LanguageProvider
+import okhttp3.Interceptor
+import okhttp3.Response
+
+class LanguageInterceptor(
+    private val languageProvider: LanguageProvider
+) : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val original = chain.request()
+
+        val ignoreLanguage = original.header("Ignore-Language") != null
+        val newRequest = original.newBuilder().removeHeader("Ignore-Language").build()
+
+        if (ignoreLanguage) {
+            return chain.proceed(newRequest)
+        }
+
+        val originalUrl = original.url
+
+        val newUrl = originalUrl.newBuilder()
+            .addQueryParameter("language", languageProvider.getCurrentLanguage()).build()
+
+        return chain.proceed(
+            newRequest.newBuilder().url(newUrl).build()
+        )
+    }
+}
